requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.3
itsdangerous==2.2.0
click==8.1.7
blinker==1.6.2

# Core ML stack
transformers==4.54.0
accelerate==1.7.0
tokenizers==0.21.1
sentencepiece==0.2.0
safetensors==0.5.3
huggingface-hub==0.34.2

# Backends
torch==2.7.0
torchvision==0.22.0
llama-cpp-python==0.3.15
pytorch-crf==0.7.2

# OCR / imaging / PDF
PyMuPDF==1.25.5  # fitz for PDF processing
Pillow==11.2.1

# Utilities
python-dotenv==1.1.1
requests==2.32.4
protobuf==6.31.1
psutil==6.1.0